fix: corrected jpa realization

+++ scrapper/src/main/java/edu/java/service/jpa/JpaSiteService.java
@@ -1,24 +1,32 @@
 package edu.java.service.jpa;
 
+import edu.java.exception.SiteAlreadyExistsException;
+import edu.java.exception.SiteNotFoundException;
 import edu.java.model.Site;
 import edu.java.repository.jpa.JpaSiteRepository;
 import edu.java.service.SiteService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
 import java.util.List;
+import lombok.RequiredArgsConstructor;
 
-@Service
 @RequiredArgsConstructor
 public class JpaSiteService implements SiteService {
     private final JpaSiteRepository repository;
 
     @Override
     public void addSite(String name) {
-        repository.save(Site.builder().name(name).build());
+        if (repository.findBySiteName(name).isPresent()) {
+            throw new SiteAlreadyExistsException("Site with URL = %s already exists".formatted(name));
+        }
+
+        repository.save(Site.builder().siteName(name).build());
     }
 
     @Override
     public void removeSite(Long id) {
+        if (repository.findById(id).isPresent()) {
+            throw new SiteNotFoundException("Site not found.");
+        }
+
         repository.deleteById(id);
     }
